\name{plotAndPlayGTK-internal}
\alias{.plotAndPlay_close_event}
\alias{.plotAndPlay_save_event}
\alias{.plotAndPlay_copy_event}
\alias{.plotAndPlay_print_event}
\alias{.plotAndPlay_greyscale_event}
\alias{.plotAndPlay_zoomin_event}
\alias{.plotAndPlay_zoomout_event}
\alias{.plotAndPlay_zoomfit_event}
\alias{.plotAndPlay_centre_event}
\alias{.plotAndPlay_focus_event}
\alias{.plotAndPlay_expand_event}
\alias{.plotAndPlay_zero_event}
\alias{.plotAndPlay_logscale_event}
\alias{.plotAndPlay_prevpage_event}
\alias{.plotAndPlay_nextpage_event}
\alias{.plotAndPlay_identify_event}
\alias{.plotAndPlay_identify_region_event}
\alias{.plotAndPlay_identify_setstyle_event}
\alias{.plotAndPlay_brush_event}
\alias{.plotAndPlay_brush_region_event}
\alias{.plotAndPlay_brush_drag_event}
\alias{.plotAndPlay_clear_event}
\alias{.plotAndPlay_annotate_event}
\alias{.plotAndPlay_edit_annotations_event}
\alias{.plotAndPlay_edit_in_place_event}
\alias{.plotAndPlay_edit_event}
\alias{.plotAndPlay_layers_event}
\alias{.plotAndPlay_zoomin3d_event}
\alias{.plotAndPlay_zoomout3d_event}
\alias{.plotAndPlay_flyleft3d_event}
\alias{.plotAndPlay_flyright3d_event}
\alias{.plotAndPlay_index_event}
\title{Internal plotAndPlayGTK user interface callbacks}
\description{These are exported from the package so that the GUI can
  access the callbacks. They should otherwise be ignored.}
\author{Felix Andrews \email{felix@nfrac.org}}
\keyword{internal}

