\name{plotAndPlayGTK-package}
\alias{plotAndPlayGTK-package}
\alias{plotAndPlayGTK}
\docType{package}
\title{
plotAndPlayGTK: A GUI for interactive plots using GTK+
}
\description{
	See \code{help(package=plotAndPlayGTK)} and \code{\link{plotAndPlay}}.
}
\details{
\tabular{ll}{
Package: \tab plotAndPlayGTK\cr
Type: \tab Package\cr
Version: \tab \code{plotAndPlayGTK:::VERSION}\cr
Date: \tab 2007-06-21\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{ Felix Andrews \email{felix@nfrac.org} }
\keyword{ package }
\seealso{
	\code{\link{plotAndPlay}}
}

